Promote bool and char to int in function paramters
This makes it work to for example declare C functions:

    isalpha(char) bool
    isdigit(char) bool

Which gets translated to:

    isalpha(int) int
    isdigit(int) int

Which makes things easier since one doesn't have to explicitly compare
the return value of the functions with 0 to use in an if-condition.

This should be safe for function parameters, since C promotes all
integer types smaller than int to int.  Not completely sure about return
types though.