Only have for for loops and remove do and do while
- "for { ... }" is equivalent to "while true { ... }"
- "for expr { ... }" is equivalent to "while expr { ... }"

By the same rules, this also makes:

- "if { ... }" is equivalent to "if true { ... }"

Which can be used to implement compound statements previously written
using the "do" keyword.

This also removes the "do while" loop completely, as "do for" sounds a
bit uncommon.  This probably isn't a problem because it can be written
using the "break" and "if" statements instead.