Make semicolon a statement terminator
Previously it was statement delimeter so one could write things like `if
expr { return }`.  Now every statement (except for those ending in
blocks) has to be terminated with a semicolon, so it is required to
write `if expr { return; }`.  This change makes it so that structures
can more easily be disambiguated from blocks which simplifies the
parser.

Another change is that the BCPL semicolon insertion rules at newlines in
the lexer have been removed.  These rules where put there because the
grammar previously required semicolons at the end of every statement
including `if`, `for` and `switch` which made the language quite
verbose.  Removing these semicolon insertion rules makes the language
free-form like C.